+++ hmf-common/src/main/java/com/hartwig/hmftools/common/copynumber/CopyNumberAlteration.java
@@ -10,7 +10,7 @@ public enum CopyNumberAlteration {
     private static final int NORMAL_HUMAN_COPY_NUMBER = 2;
 
     @NotNull
-    public static CopyNumberAlteration fromCopyNumber(final double value) {
+    public static CopyNumberAlteration fromCopyNumber(final int value) {
         if (value == NORMAL_HUMAN_COPY_NUMBER) {
             return NEUTRAL;
         }
